Implement CRI detection for Windows
