+++ cmd/kubeadm/app/apis/kubeadm/v1beta2/defaults_windows.go
@@ -22,5 +22,5 @@ const (
 	// DefaultCACertPath defines default location of CA certificate on Windows
 	DefaultCACertPath = "C:/etc/kubernetes/pki/ca.crt"
 	// DefaultUrlScheme defines default socket url prefix
-	DefaultUrlScheme = "tcp"
+	DefaultUrlScheme = "npipe"
 )
