+++ cmd/kubeadm/app/constants/constants_windows.go
@@ -20,5 +20,5 @@ package constants
 
 const (
 	// DefaultDockerCRISocket defines the default Docker CRI socket
-	DefaultDockerCRISocket = "tcp://localhost:2375"
+	DefaultDockerCRISocket = "npipe:////./pipe/docker_engine"
 )
