+++ cmd/kubeadm/app/util/runtime/runtime_unix.go
@@ -1,7 +1,7 @@
-// +build windows
+// +build !windows
 
 /*
-Copyright 2018 The Kubernetes Authors.
+Copyright 2019 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -16,11 +16,23 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package v1beta1
+package util
+
+import (
+	"os"
+)
 
 const (
-	// DefaultCACertPath defines default location of CA certificate on Windows
-	DefaultCACertPath = "C:/etc/kubernetes/pki/ca.crt"
-	// DefaultUrlScheme defines default socket url prefix
-	DefaultUrlScheme = "npipe"
+	dockerSocket     = "/var/run/docker.sock" // The Docker socket is not CRI compatible
+	containerdSocket = "/run/containerd/containerd.sock"
 )
+
+// isExistingSocket checks if path exists and is domain socket
+func isExistingSocket(path string) bool {
+	fileInfo, err := os.Stat(path)
+	if err != nil {
+		return false
+	}
+
+	return fileInfo.Mode()&os.ModeSocket != 0
+}
